Add files via upload
This file is an exploration of data from a fictional superstore company's locations across the country by region, state and city. It provides information on the categories and subcategories of products they provide as well as their sales and profits. I used my queries here to examine the least and most profitable products the stores offer and which areas of the country that their stores perform best and worst.