Add files via upload
Tableau workbook containing the dashboard for the food macros dataset. The dashboard consists of a table that displays food name, calorie count, protein, carb and fat content. It can be filtered by diet type and a drop down list that allows the user to select different food categories to further filter the table.